Put the [multiplayer] top-level tag in editor-generated cfg files.
Currently the editor created cfg files with no top-level tag at all, which isn't valid WML and in any other situation would be treated as an error. Having the top-level tags be present then allows a UMC author to move a scenario from the editor to an add-on without needing to modify the scenario cfg file manually themselves.