Use file extensions rather than file contents to determine how to load a file in the editor.
I was surprised to see that the editor tries to determine what something is by looking at the file contents rather than relying on the long established .map and .cfg conventions, though I suspect this was also the original reason behind the editor generated scenario files not having a top-level tag. In any case, this commit changes it to the following:
* .map - pure map file
* .ed.cfg - editor generated scenario file
* .cfg - non-editor generated scenario file

Also use std::string's find and substr methods instead of regex, thus resolving one use of match_dot_not_null by just not using regex in the first place.