campaignd: Refactor [request_campaign]
This makes the code easier to follow, hopefully, and quite importantly
avoids an unnecessary copy while building the update pack sequence.

It does away with the size report for deltas in stderr because it's
inaccurate by design -- it's the sum of the individual packs' compressed
sizes rather than the size of the compressed result (which we cannot
determine in advance.)

While testing this I ran into an issue where apparently someone didn't
implement the from/to version logic for the update pack sequence, who
knew. Gotta fix that later I reckon.