campaignd/fs_commit: Reimplement for Windows
This replaces the existing low-effort implementation on Windows (which
was broken and made it impossible to use campaignd in a meaningful
fashion at all) with one that actually attempts to replace files in the
most atomic fashion we can afford using the Windows API.

Temporary files are now opened with FILE_SHARE_DELETE so that we can
perform rename and delete operations on them without having to close
them first, which would be an open invitation for UB if any references
held to the underlying ostream get used again.

Instead of calling std::remove and std::rename (both of which weren't
Unicode-safe anyway) we now use the kernel32 API
SetFileInformationByHandle() to forcefully replace a potentially
existing file. It's a win-win.

Some of the POSIX code is touched here in order to keep things tidy;
namely the get_stream_file_descriptor() helper.