GUI2/Drop Down Menu: handle toggle callbacks using the event system
This fires a NOTIFIED_MOFIDIED event in the parent widget when the toggle state changes instead of manually
passing a callback in as a ctor parameter. The only widget which handled this was the multimenu_button, and
that did some stuff, then fired off a NOTIFY_MODIFIED event of its own. This change basically just installs
a handler of its own for the NOTIFY_MODIFIED event, which should still allow any other dialogs to add their
own handlers if necessary.

This also adds a second constructor that takes the parent widget and pulls the rectangle and markup settings
from it rather than having them passed in manually. The ctor that takes them is retained for the cases where
the menu needs to be manually invoked without a parent widget.