DiD S9: small fixes
Some units can be cut of fin the 2nd round. That's not a problem, but with a
good portion of bad luck it could happen to MKBones, making the scenario
unwinnable. That's handled by map adjustments.

Also increasing HP by 3 for the lich start, might even be more.
Discussion in #5296