wml/preprocessor: Reduce code duplication in preprocessor conditionals
The code handling #if(n)def and #if(n)have is largely the same in both
the normal and the negated paths, making two code paths redundant. This
commit unifies each pair as it should have been done from the beginning.

This commit also alters the implementation of #if(n)ver very slightly so
we do a single char check instead of comparing against a whole string a
second time (we definitely don't need to compare the full thing).