campaignd: Add server info request, [server_id]
This new commands allows clients to request some identification and
feature capability information from the server, intended to allow
clients to find out what features the server supports and enable or
disable UI elements accordingly while preserving compatibility where
desirable.

Currently this results in a response like this:

  [server_id]
      version="1.15.7+dev (c7643e0785b-Modified)"
      cap="auth:legacy,delta"
  [/server_id]

More information might be added later, such as a website URL and a link
to a web listing for the particular server (e.g. <https://add-ons.wesnoth.org/1.16/>),
more details on add-on licensing, or even doing away with the need to
send [request_terms] before displaying the prompt on the client side.