addon/client: Initial support for [server_id]
The client now sends the [server_id] command during connection and
records the server version and capabilities for later use. A simple API
to query the client's saved info is provided but not used at this time.

If the server does not recognize [server_id], the request will result in
an error that we simply ignore, and we make an educated guess on
capabilities based on vanilla campaignd.