fs: Un-inline non-trivial blacklist_pattern_list methods
This allows us to drop the hard dependency on <algorithm>, <functional>,
and serialization/string_utils.hpp from filesystem.hpp, which should
make the latter less of a heavyweight include in places that otherwise
don't need those headers.

There are no performance concerns here because the class as a whole is
rarely used in general and never in performance-sensitive code.

This commit also addresses a few side-effects of other files
piggybacking on the filesystem.hpp dependencies.