campaignd: Add licensing info and add-ons web URL to [server_id]
This is the part where an attentive reader realises the [server_info]
tag was intended all along to contain information for the command's
response, not just the add-on thread URL formatting string. Talk about a
Chekhov's gun.

Still want to keep them named differently since [server_id] is a
superset of [server_info] including hardcoded values (capabilities and
version).