campaignd: Add strict version check on uploads
This makes it so by default campaignd will reject uploads of existing
add-ons without a version bump. A server-side configuration flag
(strict_versions) can be optionally enabled to lower the strictness and
allow re-uploading the same version -- while still disallowing going
backwards.

The reason for making it possible to lower the strictness is that very
often while testing campaignd itself it becomes necessary to reupload
the same add-on over and over. We don't want to get rid of *that*
particular use case.

The server does not currently have a mechanism to advertise the
different strictness levels, unfortunately, so clients may in fact need
to send a 100 MB rock before finding out it's not good enough for
upload. Such is life. This might be solved after #5338 is merged into
mainline.

Closes #5079.