Laid groundwork for unified network data processing (instead of dialog-based)
Currently, only the initial gamelist is handled by this. The idea is to treat that list like any other,
as opposed to having special handling in open_connection. This is since there may be other server data
received prior to the initial gamelist ([message]s for instance) that we'll want to handle without
adding special handling in open_connection or in some other fashion.