addon/mg: Refactor installed/available_addons()
The code in both is almost identical except for the check for .pbl
files. Furthermore, the call to filesystem::get_files_in_dir() does not
need a list of files to output since support for Addon_Name.cfg-style
add-ons was dropped in version 1.13.0.

The ADDON_ENUM_CRITERIA enumeration is there in case in the future we
need more criteria for enumerating add-ons. Right now it's just a
convenience so enumerate_addons_internal() takes a self-descriptive
parameter instead of a boolean.