Tell other clients about controller type changes from :droid
If one client changes the controller type of a side it controls from human -> AI or vice versa, the other clients need to know about it in case there's WML/lua that uses the controller type to determine what to do. This also means that droiding a side you own can now only be done when it's your turn.

Fixes #4996