GUI2/Label: properly overload update_canvas
Previously, I was handling updating the label alpha in the set_text_alpha function, which meant it needed
to be called from the builder in order for that variable to be properly set on a newly constructed label.
This properly adds it to the update_canvas and changes set_text_alpha to match the other setter formats
that require canvas updates.