gui2/campaign_selection: Add difficulty-specific laurels
This brings back laurels for the difficulty menu. Because I really don't
want to break into menu_button's implementations details just yet, this
is a bit of a hack using the BLIT IPF instead of stacked_widget.

To make up for it, this version uses the three different difficulty
laurels depending on the difficulty being displayed, instead of always
using the standard laurel.