Add help topics for the Editor, as it stands in the 1.14 branch
This commit is on the master branch. There have been no changes at all to the
editor documentation in the 1.14 branch, while there have been several updates
to the master branch that would be good to backport if the help is displayed in
the 1.14 branch - therefore my plan is to get this set of changes reviewed in
the master branch, and then simply pull the entire file into the 1.14 branch
along with enabling the editor documentation (a small change to
data/core/help.cfg, as done in commit
ed611f94385f608512da55d691d3c9753240d5d5).

A couple of warnings about the scenario mode being buggy are added to the
user-visible text, find these by looking for <bold> tags within the text.