Wesnothd: cleaned up usage of document smart pointers
This removes the use of boost::ptr_vector and replaces it with a vector of unique_ptr<document>.
document::clone was also made to return a unique_ptr, which allows for the unification of doc
pointer management. We don't have stray document pointers anymore, there's no manual deletion,
and it's clear who owns what when (ptr_vector maintains ownership of the pointers it holds, but
it's clearer to have everything using unique_ptr).