Refactored out preferences::is_authenticated and friends
This moves all the is-autenticated (really is-moderator) stuff out of preferences and into a new session_metadata class.
Note this does remove the ability to [de]auth as a mod while logged in for now.

This also makes open_connection a member of mp_manager and splits some implementations out from their declarations.