Re-enable the editor help, add docs about keyboard modifiers and the scenario editor
A couple of warnings about the scenario mode being buggy are added to the
user-visible text, find these by looking for <bold> tags within the text.

This takes all the changes to the editor documentation that were made to master
up to and including a4711ebad9. There have been no changes at all to the editor
documentation in the 1.14 branch, while there have been several updates to the
master branch that would be good to backport if the help is displayed in the
1.14 branch - this commit simply pulls the entire file into the 1.14 branch
along with enabling the editor documentation (a small change to
data/core/help.cfg, as done in commit ed611f9438).

The sound source tool isn't mentioned at all in this documentation. There is a
button for it in the 1.14 UI, but that button's only function is to say that
it's not implemented, so documentation seems unnecessary.