Game Config Manager: removed jump_to_editor ctor argument
This was taken from the value in game_launcher, which was set to true if commandline_options::editor was set.
This uses that value instead of passing in the flag manually, which is consistent with the other defines.

This also removes game_launcher::jump_to_editor since it's no longer needed.