[wmllint] DiD: suppress: Inky speaks in his/her "die" event rather than "last breath"
It is a common mistake to have a unit speak in his/her "die" event (after the death animation is played). Wmllint tries to catch that. In this case, however, Inky does not actually die (as far as the plot is concerned) so it is a perfectly
valid stylistic choice for her to make a "Bloub" sound as she dives into the water.