Always have units with multiple advancements level up to the first one listed in the unit_type.
There's already enough randomness in Wesnoth; having the game randomly choose what a unit levels up to in a multiplayer game because it happens to advance on someone else's turn is unnecessary.
This also reduces the need to manually select a unit's advancement path for advancement planning modifications. For example, if you want your Mage to level up into a White Mage, this is now the default and doesn't need to be manually selected by the player.