campaignd: Fix UB when servicing [request_campaign]
std::distance(A, B) requires A and B to be random access iterators if
the expression B < A is even expected to make any sense. std::map
produces bidirectional iterators instead, so this code was causing UB
when [request_campaign] was used with a From version newer than the To
version (a.k.a. downgrading an add-on).

This fix also allows us to skip a lot of nonsense in the event that From
== To.