hotkey: Streamline some menu action labels
Something that people occasionally miss is that several hotkey actions
double as menu items throughout the UI. Menu items have this particular
quality to them that requires them to be as succinct as possible to make
their meaning apparent without requiring excessive reading and parsing
from users/players. Furthermore, in a UI like ours with large elements
everywhere,  we want to avoid excessively wordy elements that can become
ten times as long when translated into foreign languages.

Commit c42bf114249dcef5241327ca92214cf155ff89e8 replaced the Preferences
label with a unnecessarily longer version despite it being pretty
standard among most applications for it to be a single word (either
Preferences or Options or Settings). This commit reverts that.

Clear Chat and Errors (also from the aforementioned commit) is relabeled
Clear Chat rather than Clear Messages so it's clearer that it applies to
chat messages in particular and not anything else. On that note, the
fact that game diagnostics share real estate and infrastructure with
chat is really unfortunate but there's not much that can be done about
it at the moment. At least it is visually obvious that this is the case.
It's also rarely relevant (and most people won't think of entering
"error" into the hotkey filter before reporting the error somewhere).

View Chat Log is relabeled Chat Log since no other option deals with the
Chat *Log* in particular.

Scenario Objectives is relabeled Objectives because there really isn't
anything else that has objectives we can look at in the whole game.

Enter User Command is relabeled Enter Command so as to not sound
superfluous.

Additionally, the Quit to Main Menu action is relabeled just Quit to
Menu, which is closer to what the option actually does in MP games (quit
to lobby) without being inaccurate in either SP or MP (closes #2744).