Add [test] to list of tags with add-on id
As of PR #4963, the engine ignores all global tags defined by add-ons that are not active, that is, that are not used in the current game.  Whether an add-on counts as active is determined based on a list of possible tags with add-on ids.  The [test] tag is not included in this list, resulting in test scenarios in add-ons not working any more.